requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pytest==8.3.5
google-api-python-client==2.165.0
google-auth==2.38.0

google-auth-oauthlib==1.2.1
